+++ src/Components/Intro.tsx
@@ -1,11 +1,11 @@
-import project1 from "../Images/에코초이스.webp";
-import project2 from "../Images/솜뭉치.webp";
-import project3 from "../Images/뜨개랑.webp";
+import project1 from "../Images/2024/에코초이스.webp";
+import project2 from "../Images/2024/솜뭉치.webp";
+import project3 from "../Images/2024/뜨개랑.webp";
 
-import Introduction1 from "../Images/1.webp";
-import Introduction2 from "../Images/2.webp";
-import Introduction3 from "../Images/3.webp";
-import Introduction4 from "../Images/4.webp";
+import Introduction1 from "../Images/2024/1.webp";
+import Introduction2 from "../Images/2024/2.webp";
+import Introduction3 from "../Images/2024/3.webp";
+import Introduction4 from "../Images/2024/4.webp";
 
 import ProjectCard from "./ProjectCard";
