+++ src/Pages/AdminPage.tsx
@@ -11,8 +11,8 @@ import {
   orderBy,
   addDoc,
 } from "firebase/firestore";
-import Footer from "../Components/Footer";
-import NavBar from "../Components/NavBar";
+import Footer from "../Components/common/Footer";
+import NavBar from "../Components/common/NavBar";
 import { adminApp, db } from "../firebase";
 import { useEffect, useState } from "react";
 import RequestTable from "../Components/RequestTable";
