+++ src/Pages/MainPage.tsx
@@ -1,7 +1,7 @@
-import NavBar from "../Components/NavBar";
+import NavBar from "../Components/common/NavBar";
 import Intro from "../Components/Intro";
-import Footer from "../Components/Footer";
-import FAQ from "../Components/FAQ";
+import Footer from "../Components/common/Footer";
+import FAQ from "../Components/common/FAQ";
 const MainPage = () => {
   return (
     <section className="mainpage">
