+++ src/Pages/ProjectPage.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useRef, useState } from "react";
-import NavBar from "../Components/NavBar";
+import NavBar from "../Components/common/NavBar";
 import ProjectCard from "../Components/ProjectCard";
-import Footer from "../Components/Footer";
+import Footer from "../Components/common/Footer";
 import {
   QueryDocumentSnapshot,
   QueryFieldFilterConstraint,
@@ -23,7 +23,7 @@ import useStore from "../store";
 import { HiPencilSquare } from "react-icons/hi2";
 import Pagination from "../Components/Pagination";
 import LoadingCircle from "../Components/LoadingCircle";
-import ScrollToTop from "../Components/ScrollToTop";
+import ScrollToTop from "../Components/common/ScrollToTop";
 
 type Filter = "default" | "app" | "web" | "personal" | "team";
 type MyIndexType = {
